+++ tests/Rule/GuessableDataRuleTest.php
@@ -12,21 +12,21 @@ use Stadly\PasswordPolice\WordFormatter;
 use Traversable;
 
 /**
- * @coversDefaultClass \Stadly\PasswordPolice\Rule\GuessableData
+ * @coversDefaultClass \Stadly\PasswordPolice\Rule\GuessableDataRule
  * @covers ::<protected>
  * @covers ::<private>
  */
-final class GuessableDataTest extends TestCase
+final class GuessableDataRuleTest extends TestCase
 {
     /**
      * @covers ::__construct
      */
     public function testCanConstructRule(): void
     {
-        $rule = new GuessableData();
+        $rule = new GuessableDataRule();
 
         // Force generation of code coverage
-        $ruleConstruct = new GuessableData();
+        $ruleConstruct = new GuessableDataRule();
         self::assertEquals($rule, $ruleConstruct);
     }
 
@@ -35,7 +35,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testRuleIsSatisfiedWhenPasswordIsString(): void
     {
-        $rule = new GuessableData();
+        $rule = new GuessableDataRule();
 
         self::assertTrue($rule->test('foobar'));
     }
@@ -45,7 +45,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testPasswordCanContainGuessableStringInRule(): void
     {
-        $rule = new GuessableData(['oba']);
+        $rule = new GuessableDataRule(['oba']);
         $password = new Password('foobar');
 
         self::assertFalse($rule->test($password));
@@ -56,7 +56,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testPasswordCanContainGuessableStringInPassword(): void
     {
-        $rule = new GuessableData();
+        $rule = new GuessableDataRule();
         $password = new Password('foobar', ['oba']);
 
         self::assertFalse($rule->test($password));
@@ -67,7 +67,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testPasswordCanNotContainGuessableString(): void
     {
-        $rule = new GuessableData(['oob']);
+        $rule = new GuessableDataRule(['oob']);
         $password = new Password('foo28/11/18bar', ['oba']);
 
         self::assertTrue($rule->test($password));
@@ -78,7 +78,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testPasswordCanContainGuessableDateInRule(): void
     {
-        $rule = new GuessableData([new DateTime('2018-11-28')]);
+        $rule = new GuessableDataRule([new DateTime('2018-11-28')]);
         $password = new Password('foo28/11/18bar');
 
         self::assertFalse($rule->test($password));
@@ -89,7 +89,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testPasswordCanContainGuessableDateInPassword(): void
     {
-        $rule = new GuessableData();
+        $rule = new GuessableDataRule();
         $password = new Password('foo28/11/18bar', [new DateTime('2018-11-28')]);
 
         self::assertFalse($rule->test($password));
@@ -100,7 +100,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testPasswordCanNotContainGuessableDate(): void
     {
-        $rule = new GuessableData();
+        $rule = new GuessableDataRule();
         $password = new Password('foobar', [new DateTime('2018-11-28')]);
 
         self::assertTrue($rule->test($password));
@@ -111,7 +111,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testRuleIsSatisfiedWhenConstraintWeightIsLowerThanTestWeight(): void
     {
-        $rule = new GuessableData([], [], 1);
+        $rule = new GuessableDataRule([], [], 1);
         $password = new Password('foobar', ['oba']);
 
         self::assertTrue($rule->test($password, 2));
@@ -131,7 +131,7 @@ final class GuessableDataTest extends TestCase
             }
         );
 
-        $rule = new GuessableData([], [$wordFormatter]);
+        $rule = new GuessableDataRule([], [$wordFormatter]);
 
         self::assertFalse($rule->test(new Password('pine4ppl€jack', ['apple'])));
         self::assertTrue($rule->test(new Password('pine4pp1€jack', ['apple'])));
@@ -151,7 +151,7 @@ final class GuessableDataTest extends TestCase
             }
         );
 
-        $rule = new GuessableData([], [$wordFormatter]);
+        $rule = new GuessableDataRule([], [$wordFormatter]);
 
         self::assertFalse($rule->test(new Password('foo2B/II/1Bbar', [new DateTime('2018-11-28')])));
         self::assertTrue($rule->test(new Password('fooZB/I!/1Bbar', [new DateTime('2018-11-28')])));
@@ -180,7 +180,7 @@ final class GuessableDataTest extends TestCase
             }
         );
 
-        $rule = new GuessableData([], [$wordFormatter1, $wordFormatter2]);
+        $rule = new GuessableDataRule([], [$wordFormatter1, $wordFormatter2]);
 
         self::assertTrue($rule->test(new Password('pine4ppl€jack', ['apple'])));
         self::assertFalse($rule->test(new Password('pineappl€jack', ['apple'])));
@@ -210,7 +210,7 @@ final class GuessableDataTest extends TestCase
             }
         );
 
-        $rule = new GuessableData([], [$wordFormatter1, $wordFormatter2]);
+        $rule = new GuessableDataRule([], [$wordFormatter1, $wordFormatter2]);
 
         self::assertTrue($rule->test(new Password('foo2B/I!/1Bbar', [new DateTime('2018-11-28')])));
         self::assertFalse($rule->test(new Password('foo28/II/18bar', [new DateTime('2018-11-28')])));
@@ -222,7 +222,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testRuleCanBeValidated(): void
     {
-        $rule = new GuessableData();
+        $rule = new GuessableDataRule();
         $password = new Password('test', ['oba', new DateTime('2018-11-28')]);
 
         self::assertNull($rule->validate($password));
@@ -233,7 +233,7 @@ final class GuessableDataTest extends TestCase
      */
     public function testRuleCanBeInvalidated(): void
     {
-        $rule = new GuessableData();
+        $rule = new GuessableDataRule();
         $password = new Password('foobar', ['oba', new DateTime('2018-11-28')]);
 
         self::assertEquals(