+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Renamed `CharacterClass` rule to `CharacterClassRule`.
 - Renamed `Dictionary` rule to `DictionaryRule`.
 - Renamed `Digit` rule to `DigitRule`.
+- Renamed `GuessableData` rule to `GuessableDataRule`.
 
 ### Fixed
 - Nothing